refactor(app): split game_map_area part1 for part-unit size lint
Move feed entries, didUpdateWidget, and region viewport snapshot helpers
into game_map_area_part1b.dart so game_map_area_part1.dart stays within
the 1000-line fragment cap (repo.part_unit_size).
